+++ docs/.vitepress/config.ts
@@ -46,9 +46,22 @@ export default defineConfig({
               text: 'Quick Start',
               link: '/guide/basics/quick-start',
             },
+          ],
+        },
+        {
+          text: 'Contributing',
+          items: [
             {
               text: 'Component Creation',
-              link: '/guide/basics/component-creation',
+              link: '/guide/contributing/component-creation',
+            },
+            {
+              text: 'Component Documentation',
+              link: '/guide/contributing/component-documentation',
+            },
+            {
+              text: 'Contribution',
+              link: '/guide/contributing/contribution',
             },
           ],
         },
@@ -62,32 +75,60 @@ export default defineConfig({
           text: 'Components',
           items: [
             {
-              text: 'Button',
-              link: '/documentation/components/button',
-            },
-            {
-              text: 'Sidenav',
-              link: '/documentation/components/sidenav',
-            },
-            {
-              text: 'Lozenge',
-              link: '/documentation/components/lozenge',
-            },
-            {
-              text: 'Switch',
-              link: '/documentation/components/switch',
+              text: 'Avatar',
+              link: '/documentation/components/avatar',
             },
             {
               text: 'Badge',
               link: '/documentation/components/badge',
             },
             {
-              text: 'Radio',
-              link: '/documentation/components/radio',
+              text: 'Button',
+              link: '/documentation/components/button',
+            },
+            {
+              text: 'Checkbox',
+              link: '/documentation/components/checkbox',
+            },
+            {
+              text: 'Dropdown',
+              link: '/documentation/components/dropdown',
+            },
+            {
+              text: 'Empty State',
+              link: '/documentation/components/empty-state',
             },
             {
               text: 'Input',
               link: '/documentation/components/input',
+            },
+            {
+              text: 'Lozenge',
+              link: '/documentation/components/lozenge',
+            },
+            {
+              text: 'Modal',
+              link: '/documentation/components/modal',
+            },
+            {
+              text: 'Radio',
+              link: '/documentation/components/radio',
+            },
+            {
+              text: 'Sidenav',
+              link: '/documentation/components/sidenav',
+            },
+            {
+              text: 'Snackbar',
+              link: '/documentation/components/snackbar',
+            },
+            {
+              text: 'Switch',
+              link: '/documentation/components/switch',
+            },
+            {
+              text: 'Table',
+              link: '/documentation/components/Table',
             },
             {
               text: 'Tabs',
@@ -102,18 +143,6 @@ export default defineConfig({
               link: '/documentation/components/tooltip',
             },
             {
-              text: 'Modal',
-              link: '/documentation/components/modal',
-            },
-            {
-              text: 'Empty State',
-              link: '/documentation/components/empty-state',
-            },
-            {
-              text: 'Avatar',
-              link: '/documentation/components/avatar',
-            },
-            {
               text: 'Sidepanel',
               link: '/documentation/components/sidepanel',
             },
@@ -123,24 +152,24 @@ export default defineConfig({
           text: 'Utilities',
           items: [
             {
+              text: 'Border Radius',
+              link: '/documentation/utilities/border-radius',
+            },
+            {
               text: 'Colors',
               link: '/documentation/utilities/colors',
-            },
-            {
-              text: 'Typography',
-              link: '/documentation/utilities/typography',
-            },
-            {
-              text: 'Spacing',
-              link: '/documentation/utilities/spacing',
             },
             {
               text: 'Max Width',
               link: '/documentation/utilities/max-width',
             },
             {
-              text: 'Border Radius',
-              link: '/documentation/utilities/border-radius',
+              text: 'Spacing',
+              link: '/documentation/utilities/spacing',
+            },
+            {
+              text: 'Typography',
+              link: '/documentation/utilities/typography',
             },
           ],
         },