+++ src/components/tabs/tabs.vue
@@ -12,40 +12,53 @@
           'rounded-r-md': !underlined && index === tabElements.length - 1,
 
           // Regular Tab - Active
-          'text-color-brand-base border-color-success-base cursor-pointer border border-solid':
-            !underlined && activeTab.index === index,
+          'border-color-success-base cursor-pointer !border !border-solid': !underlined && activeTab.index === index,
 
           // Regular Tab - Inactive
-          'border-color-weak hover:background-color-hover cursor-pointer border-x-[0.5px] border-y border-solid':
+          'border-color-weak hover:background-color-hover cursor-pointer !border-x-[0.5px] !border-y !border-solid':
             !underlined && activeTab.index !== index,
 
           // Regular Tab - Disabled
-          'background-color-base !cursor-not-allowed': !underlined && tab.disabled,
+          'background-color-disabled !cursor-not-allowed': !underlined && tab.disabled,
 
           // Underlined Tab - Active
-          'text-color-brand-base cursor-pointer': underlined && activeTab.index === index,
+          'cursor-pointer': underlined && activeTab.index === index,
 
           // Underlined Tab - Inactive
-          'border-color-base cursor-pointer border-b border-solid': underlined && activeTab.index !== index,
+          'border-color-base cursor-pointer !border-b !border-solid': underlined && activeTab.index !== index,
           'hover:background-color-hover cursor-pointer': underlined && activeTab.index !== index && !tab.disabled,
 
           // Underlined Tab - Disabled
-          'border-color-disabled text-color-disabled !cursor-not-allowed border-b': underlined && tab.disabled,
+          'border-color-disabled text-color-on-fill-disabled background-color-disabled !cursor-not-allowed border-b':
+            underlined && tab.disabled,
         },
       ]"
       @click="updateSelectedTabIndex(index, tab.disabled)"
     >
+      <!-- Regular Tab Background Active Indicator -->
+      <div
+        v-if="!underlined && activeTab.index === index"
+        :class="[
+          'background-color-single-active tw-w-full absolute bottom-0 left-0 z-[5] block h-full w-full',
+          {
+            'rounded-l-md': activeTab.index === 0,
+            'rounded-r-md': activeTab.index === tabElements.length - 1,
+          },
+        ]"
+      />
+
       <div
         :class="{
-          'flex items-center gap-size-spacing-5xs leading-none': true,
-          'transition duration-150 ease-in-out': true,
-          'cursor-pointer group-active:scale-75': !tab.disabled,
+          'relative z-[10] flex items-center gap-size-spacing-5xs leading-none': true,
           'cursor-not-allowed': tab.disabled,
         }"
       >
         <div v-if="!!tab.icon">
           <Icon
-            class="body-sm-regular"
+            :class="{
+              'body-sm-regular': true,
+              'text-color-brand-base': activeTab.index === index,
+            }"
             :icon="activeTab.index === index && typeof tab.iconFill === 'string' ? tab.iconFill : tab.icon"
           />
         </div>
@@ -54,24 +67,6 @@
         </div>
       </div>
     </div>
-
-    <!-- Regular Tab Background Active Indicator -->
-    <div
-      v-if="!underlined"
-      :class="[
-        'background-color-single-active absolute bottom-0 left-0 z-[5] block',
-        'transition-left duration-150 ease-in-out',
-        {
-          'rounded-l-md': activeTab.index === 0,
-          'rounded-r-md': activeTab.index === tabElements.length - 1,
-        },
-      ]"
-      :style="{
-        height: `${activeTab.height}px`,
-        width: `${activeTab.width}px`,
-        left: `${activeTab.undelineLeftOffset}px`,
-      }"
-    />
 
     <!-- Underlined Tab Active Indicator -->
     <div
@@ -82,7 +77,7 @@
       ]"
       :style="{
         width: `${activeTab.width}px`,
-        left: `calc(${activeTab.undelineLeftOffset}px - 2.3px)`,
+        left: `${activeTab.undelineLeftOffset}px`,
       }"
     />
   </div>